updated data-loader and try to fix issue with flush on loop

+++ libraries/data-logger/data_logger_config.h
@@ -12,7 +12,7 @@ typedef struct
     bool force_full_flush_erase;
     
     uint8_t type;
-    size_t entry_size;
+    uint16_t entry_size;
 } DataLoggerConfig;
 
 #endif // _DATA_LOGGER_CONFIG_H
